# docs/content/lobby/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Extinction lobby server API
  version: 0.1.0-dev

paths:
  /p/auth/sign-up:
    post:
      summary: Create a new user registration.
      description: This endpoint creates a new user registration.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: >-
                    A Unicode string with length between 6 and 30 bytes when
                    encoded in UTF-8.
                user:
                  $ref: "#/components/schemas/user"
      responses:
        "200":
          description: Registration was successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/token-response"
        "409":
          description: >-
            Registration was not successful because the username is already
            taken.

  /p/auth/sign-in:
    post:
      summary: Retrieve a fresh JWT.
      description: >-
        This endpoint is used to retrieve a fresh JWT to be used for
        authentication to other API endpoints.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string

      responses:
        "200":
          description: The username exists and the password is correct.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/token-response"
        "401":
          description: >-
            The username does not exist or the password is not correct.
  /a/games:
    post:
      summary: Create and join a new game.
      description: >-
        This endpoint is used to crate a new game. The user automatically joins
        the game. It is not possible to simultaneously be part of multiple
        games.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/game-config"
      responses:
        "200":
          description: The game was successfully crated and the user joined it.
        "400":
          description: >-
            The game cannot be created because the request is invalid. This
            might be due to invalid game configuration.
        "403":
          description: The user is already part of another game.
        "409":
          description: A different game with the same name already exists.

components:
  schemas:
    token-response:
      type: object
      properties:
        token:
          type: string
          description: >-
            A JWT token. The user uses the token to authenticate to the API.
    user:
      type: object
      properties:
        username:
          type: string
          description: >-
            A unique user name. It is a non-empty Unicode string with maximum
            length of 32 bytes when encoded in UTF-8. It does not start or end
            with whitespace.
    game-config:
      type: object
      properties:
        name:
          type: string
          description: A unique game name.
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 4
        map_name:
          type: string
          description: Name of the game map.